reposotory pattern redone
